requirements.txt
fastapi
uvicorn[standard]
google-cloud-documentai
python-multipart
